% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_factors_for_stroke_sankey_maker.R
\name{risk_factors_for_stroke_sankey_diagram}
\alias{risk_factors_for_stroke_sankey_diagram}
\title{create a sankey diagram.}
\usage{
risk_factors_for_stroke_sankey_diagram(data_to_be_plotted1,
  plot_title_in_quotes1, horizontal_spaces = TRUE,
  vertical_spaces = TRUE)
}
\arguments{
\item{data_to_be_plotted1}{a 5x6 dataframe with properly defined variables the labels row should be defined with risk_type  year1  year2 year3 year4 year5. and the risk_ type column should be defined as risk_type, Obesity, Hypertension, Diabetes, Hypercholesterolemia, and Smoking.    order of the first column does not matter, however the labels row must be in the specified order}

\item{plot_title_in_quotes1}{The preferred title of the plot in " "}

\item{horizontal_spaces}{adds horizontal spaces, must be defined as TRUE or FALSE}

\item{vertical_spaces}{adds vertical spaces, must be defined TRUE or FALSE}
}
\description{
it is only necessary to use the function, "risk_factors_for_stroke_diagram" with the properly defined parameters
takes in 5X6 dataframes with specific variables in the first column and the data for each particular year in the following columns
}
